Add 'site' foreign key relations to models
We want to be able to associate actions and 'following' with specific
sites rather than having them global.

This is a breaking change as it requires use of the site system. We
might be able to able null's in this field in order to make the support
optional. I'm not sure if that is for the best.

The django comments framework simply requires 'sites' and so does
'django-allauth' so there is some precedent for requiring it even for
people who have single sites.